requirements.txt
flask
flask_caching
waitress

gig-nuuuwan==0.1.9
